+++ src/runner/centralized_kb_hook.h
@@ -10,5 +10,4 @@ namespace CentralizedKeyboardHook
     void Stop() noexcept;
     void SetHotkeyAction(const std::wstring& moduleName, const Hotkey& hotkey, std::function<bool()>&& action) noexcept;
     void ClearModuleHotkeys(const std::wstring& moduleName) noexcept;
-    void RegisterWindow(HWND hwnd) noexcept;
 };